IE 9 Compatibillity
For IE 9 and below to display correctly, the <!DOCTYPE> declaration must be the first thing in the html file. Moved the gzip script to the second line in every files. Also added meta tags for IE compatabillity and set it to IE=edge for highest version of compatabillity available. Also added an SSI for the meta section.