Fix schedule upload permissions issue
In the old version, Apache could not  always delete the file on the server because the file that was uploaded could have restricted permissions. This changed it so that on each schedule upload, the entire schedule folder is removed and then a new one is made to place the schedule file, so apache retains permissions.